Add context struct and system-mutating funcs' unit tests

+++ download/download.go
@@ -3,30 +3,37 @@ package download
 import (
 	"fmt"
 	"github.com/yonkornilov/snowcapper/config"
+	"github.com/yonkornilov/snowcapper/context"
 	"io"
 	"net/http"
 	"os"
 )
 
-func Run(b config.Binary, target string) error {
-	fmt.Printf("Downloading %s from %s ...\n", b.Name, b.Src)
-	out, err := os.Create(target)
-	if err != nil {
-		return err
-	}
-	defer out.Close()
+func Run(c *context.Context, b config.Binary, target string) error {
+	if c.IsDryRun {
+		fmt.Printf("DRY-RUN: Downloading %s from %s ...\n", b.Name, b.Src)
+		fmt.Printf("DRY-RUN: Successfully downloaded %s to %s\n", b.Name, target)
+	} else {
+		fmt.Printf("Downloading %s from %s ...\n", b.Name, b.Src)
+		out, err := os.Create(target)
+		if err != nil {
+			return err
+		}
+		defer out.Close()
 
-	resp, err := http.Get(b.Src)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
+		resp, err := http.Get(b.Src)
+		if err != nil {
+			return err
+		}
+		defer resp.Body.Close()
 
-	_, err = io.Copy(out, resp.Body)
-	if err != nil {
-		return err
-	}
+		_, err = io.Copy(out, resp.Body)
+		if err != nil {
+			return err
+		}
 
-	fmt.Printf("Successfully downloaded %s to %s\n", b.Name, target)
+		fmt.Printf("Successfully downloaded %s to %s\n", b.Name, target)
+
+	}
 	return nil
 }
